requirements.txt
discord.py==1.7.3
fandom-py==0.2.1